+++ packages/backend-function/src/factory.test.ts
@@ -182,6 +182,17 @@ void describe('AmplifyFunctionFactory', () => {
       });
     });
 
+    void it('sets default memory', () => {
+      const lambda = defineFunction({
+        entry: './test-assets/default-lambda/handler.ts',
+      }).getInstance(getInstanceProps);
+      const template = Template.fromStack(Stack.of(lambda.resources.lambda));
+
+      template.hasResourceProperties('AWS::Lambda::Function', {
+        MemorySize: 512,
+      });
+    });
+
     void it('throws on memory below 128 MB', () => {
       assert.throws(
         () =>
@@ -335,5 +346,25 @@ void describe('AmplifyFunctionFactory', () => {
         ],
       });
     });
+  });
+
+  void it('stores single attribution data value in stack with multiple functions', () => {
+    const functionFactory = defineFunction({
+      entry: './test-assets/default-lambda/handler.ts',
+      name: 'testLambdaName',
+    });
+    const anotherFunction = defineFunction({
+      entry: './test-assets/default-lambda/handler.ts',
+      name: 'anotherName',
+    });
+    const functionStack = Stack.of(
+      functionFactory.getInstance(getInstanceProps).resources.lambda
+    );
+    anotherFunction.getInstance(getInstanceProps);
+    const template = Template.fromStack(functionStack);
+    assert.equal(
+      JSON.parse(template.toJSON().Description).stackType,
+      'function-Lambda'
+    );
   });
 });