+++ packages/backend-storage/src/storage_access_policy_factory.ts
@@ -1,8 +1,9 @@
 import { IBucket } from 'aws-cdk-lib/aws-s3';
-import { Policy, PolicyStatement } from 'aws-cdk-lib/aws-iam';
+import { Effect, Policy, PolicyStatement } from 'aws-cdk-lib/aws-iam';
 import { Stack } from 'aws-cdk-lib';
 import { AmplifyFault } from '@aws-amplify/platform-core';
 import { StorageAction, StoragePath } from './types.js';
+import { InternalStorageAction } from './private_types.js';
 
 export type Permission = {
   actions: StorageAction[];
@@ -29,7 +30,10 @@ export class StorageAccessPolicyFactory {
   }
 
   createPolicy = (
-    permissions: Readonly<Map<StorageAction, Readonly<Set<StoragePath>>>>
+    permissions: Map<
+      InternalStorageAction,
+      { allow: Set<StoragePath>; deny: Set<StoragePath> }
+    >
   ) => {
     if (permissions.size === 0) {
       throw new AmplifyFault('EmptyPolicyFault', {
@@ -39,28 +43,75 @@ export class StorageAccessPolicyFactory {
 
     const statements: PolicyStatement[] = [];
 
-    permissions.forEach((s3Prefixes, action) => {
-      statements.push(this.getStatement(s3Prefixes, action));
-    });
+    permissions.forEach(
+      ({ allow: allowPrefixes, deny: denyPrefixes }, action) => {
+        if (allowPrefixes.size > 0) {
+          statements.push(
+            this.getStatement(allowPrefixes, action, Effect.ALLOW)
+          );
+        }
+        if (denyPrefixes.size > 0) {
+          statements.push(this.getStatement(denyPrefixes, action, Effect.DENY));
+        }
+      }
+    );
+
+    if (statements.length === 0) {
+      // this could happen if the Map contained entries but all of the path sets were empty
+      throw new AmplifyFault('EmptyPolicyFault', {
+        message: 'At least one permission must be specified',
+      });
+    }
+
     return new Policy(this.stack, `${this.namePrefix}${this.policyCount++}`, {
-      statements: statements,
+      statements,
     });
   };
 
   private getStatement = (
     s3Prefixes: Readonly<Set<StoragePath>>,
-    action: StorageAction
-  ) =>
-    new PolicyStatement({
-      actions: actionMap[action],
-      resources: Array.from(s3Prefixes).map(
-        (s3Prefix) => `${this.bucket.bucketArn}${s3Prefix}`
-      ),
-    });
+    action: InternalStorageAction,
+    effect: Effect
+  ) => {
+    switch (action) {
+      case 'delete':
+      case 'get':
+      case 'write':
+        return new PolicyStatement({
+          effect,
+          actions: actionMap[action],
+          resources: Array.from(s3Prefixes).map(
+            (s3Prefix) => `${this.bucket.bucketArn}${s3Prefix}`
+          ),
+        });
+      case 'list':
+        return new PolicyStatement({
+          effect,
+          actions: actionMap[action],
+          resources: [this.bucket.bucketArn],
+          conditions: {
+            StringLike: {
+              's3:prefix': Array.from(s3Prefixes).flatMap(toConditionPrefix),
+            },
+          },
+        });
+    }
+  };
 }
 
-const actionMap: Record<StorageAction, string[]> = {
-  read: ['s3:GetObject'],
+const actionMap: Record<InternalStorageAction, string[]> = {
+  get: ['s3:GetObject'],
+  list: ['s3:ListBucket'],
   write: ['s3:PutObject'],
   delete: ['s3:DeleteObject'],
+};
+
+/**
+ * Converts a prefix like /foo/bar/* into [foo/bar/, foo/bar/*]
+ * This is necessary to grant the ability to list all objects directly in "foo/bar" and all objects under "foo/bar"
+ */
+const toConditionPrefix = (prefix: StoragePath) => {
+  const noLeadingSlash = prefix.slice(1);
+  const noTrailingWildcard = noLeadingSlash.slice(0, -1);
+  return [noLeadingSlash, noTrailingWildcard];
 };