+++ packages/create-amplify/src/get_project_root.ts
@@ -1,9 +1,8 @@
 import fsp from 'fs/promises';
 import path from 'path';
 import yargs from 'yargs';
-import { AmplifyPrompter, LogLevel } from '@aws-amplify/cli-core';
-import { printer } from './printer.js';
 import { AmplifyUserError } from '@aws-amplify/platform-core';
+import { AmplifyPrompter, LogLevel, printer } from '@aws-amplify/cli-core';
 
 /**
  * Returns the project root directory.