+++ packages/deployed-backend-client/src/deployed_backend_client_list_sandboxes.test.ts
@@ -4,7 +4,6 @@ import {
   CloudFormation,
   DescribeStacksCommand,
   ListStacksCommand,
-  ListStacksCommandInput,
   StackStatus,
 } from '@aws-sdk/client-cloudformation';
 import { BackendDeploymentStatus } from './deployed_backend_client_factory.js';
@@ -125,33 +124,37 @@ void describe('Deployed Backend Client list sandboxes', () => {
     );
   });
 
-  void it('does not paginate listSandboxes when one page contains sandboxes', async () => {
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
+  void it('does not paginate listBackends when one page contains sandboxes', async () => {
+    const sandboxes = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
+    assert.deepEqual(
+      (await sandboxes.getBackendSummaryByPage().next()).value,
+      returnedSandboxes
+    );
 
     assert.equal(listStacksMockFn.mock.callCount(), 1);
   });
 
-  void it('paginates listSandboxes when first page contains no sandboxes', async () => {
+  void it('paginates listBackends when first page contains no sandboxes', async () => {
     listStacksMockFn.mock.mockImplementationOnce(() => {
       return {
         StackSummaries: [],
         NextToken: 'abc',
       };
     });
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
+    const sandboxes = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
+    assert.deepEqual(
+      (await sandboxes.getBackendSummaryByPage().next()).value,
+      returnedSandboxes
+    );
 
     assert.equal(listStacksMockFn.mock.callCount(), 2);
   });
 
-  void it('paginates listSandboxes when one page contains sandboxes, but it gets filtered due to deleted status', async () => {
+  void it('paginates listBackends when one page contains sandboxes, but it gets filtered due to deleted status', async () => {
     listStacksMockFn.mock.mockImplementationOnce(() => {
       return {
         StackSummaries: [
@@ -162,16 +165,18 @@ void describe('Deployed Backend Client list sandboxes', () => {
         NextToken: 'abc',
       };
     });
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
+    const sandboxes = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
+    assert.deepEqual(
+      (await sandboxes.getBackendSummaryByPage().next()).value,
+      returnedSandboxes
+    );
 
     assert.equal(listStacksMockFn.mock.callCount(), 2);
   });
 
-  void it('paginates listSandboxes when one page contains sandboxes, but it gets filtered due to branch deploymentType', async () => {
+  void it('paginates listBackends when one page contains sandboxes, but it gets filtered due to branch deploymentType', async () => {
     listStacksMockFn.mock.mockImplementationOnce(() => {
       return {
         StackSummaries: [
@@ -182,16 +187,18 @@ void describe('Deployed Backend Client list sandboxes', () => {
         NextToken: 'abc',
       };
     });
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
+    const sandboxes = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
+    assert.deepEqual(
+      (await sandboxes.getBackendSummaryByPage().next()).value,
+      returnedSandboxes
+    );
 
     assert.equal(listStacksMockFn.mock.callCount(), 2);
   });
 
-  void it('paginates listSandboxes when one page contains a stack, but it gets filtered due to not having gen2 outputs', async () => {
+  void it('paginates listBackends when one page contains a stack, but it gets filtered due to not having gen2 outputs', async () => {
     getOutputMock.mock.mockImplementationOnce(() => {
       throw new BackendOutputClientError(
         BackendOutputClientErrorType.METADATA_RETRIEVAL_ERROR,
@@ -208,16 +215,18 @@ void describe('Deployed Backend Client list sandboxes', () => {
         NextToken: 'abc',
       };
     });
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
+    const sandboxes = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
+    assert.deepEqual(
+      (await sandboxes.getBackendSummaryByPage().next()).value,
+      returnedSandboxes
+    );
 
     assert.equal(listStacksMockFn.mock.callCount(), 2);
   });
 
-  void it('does not paginate listSandboxes when one page throws an unexpected error fetching gen2 outputs', async () => {
+  void it('does not paginate listBackends when one page throws an unexpected error fetching gen2 outputs', async () => {
     getOutputMock.mock.mockImplementationOnce(() => {
       throw new Error('Unexpected Error!');
     });
@@ -231,46 +240,9 @@ void describe('Deployed Backend Client list sandboxes', () => {
         NextToken: 'abc',
       };
     });
-    const listSandboxesPromise = deployedBackendClient.listSandboxes();
-    await assert.rejects(listSandboxesPromise);
-  });
-
-  void it('includes a nextToken when there are more pages', async () => {
-    listStacksMockFn.mock.mockImplementation(() => {
-      return {
-        StackSummaries: listStacksMock.StackSummaries,
-        NextToken: 'abc',
-      };
+    const listBackendsPromise = deployedBackendClient.listBackends({
+      deploymentType: 'sandbox',
     });
-    const sandboxes = await deployedBackendClient.listSandboxes();
-    assert.deepEqual(sandboxes, {
-      nextToken: 'abc',
-      sandboxes: returnedSandboxes,
-    });
-
-    assert.equal(listStacksMockFn.mock.callCount(), 1);
-  });
-
-  void it('accepts a nextToken to get the next page', async () => {
-    listStacksMockFn.mock.mockImplementation(
-      (input: ListStacksCommandInput) => {
-        if (!input.NextToken) {
-          return {
-            StackSummaries: listStacksMock.StackSummaries,
-            NextToken: 'abc',
-          };
-        }
-        return listStacksMock;
-      }
-    );
-    const sandboxes = await deployedBackendClient.listSandboxes({
-      nextToken: 'abc',
-    });
-    assert.deepEqual(sandboxes, {
-      nextToken: undefined,
-      sandboxes: returnedSandboxes,
-    });
-
-    assert.equal(listStacksMockFn.mock.callCount(), 1);
+    await assert.rejects(listBackendsPromise.getBackendSummaryByPage().next());
   });
 });