+++ packages/integration-tests/src/test-projects/data-storage-auth-with-triggers-ts/hotswap-update-files/data/resource.ts
@@ -20,12 +20,38 @@ const schema = a.schema({
     executionDuration: a.float(),
   }),
 
+  customQuery: a
+    .query()
+    .arguments({ id: a.string() })
+    .returns(a.ref('Todo'))
+    .authorization([a.allow.private()])
+    .handler(
+      // provisions JS resolver
+      a.handler.custom({
+        dataSource: a.ref('Todo'),
+        entry: './js_custom_fn.js',
+      })
+    ),
+
   echo: a
     .query()
     .arguments({ content: a.string() })
     .returns(a.ref('EchoResponse'))
     .authorization([a.allow.private()])
-    .function('echo'),
+    .handler(a.handler.function('echo')),
+
+  echoInline: a
+    .query()
+    .arguments({ content: a.string() })
+    .returns(a.ref('EchoResponse'))
+    .authorization([a.allow.private()])
+    .handler(
+      a.handler.function(
+        defineFunction({
+          entry: './echo/handler2.ts',
+        })
+      )
+    ),
 }) as never; // Not 100% sure why TS is complaining here. The error I'm getting is "The inferred type of 'schema' references an inaccessible 'unique symbol' type. A type annotation is necessary."
 
 export type Schema = ClientSchema<typeof schema>;