+++ packages/create-amplify/CHANGELOG.md
@@ -1,5 +1,30 @@
 # create-amplify
 
+## 0.7.0-beta.4
+
+### Patch Changes
+
+- ee247fd: use printer from cli-core
+- Updated dependencies [3e34244]
+- Updated dependencies [ee247fd]
+- Updated dependencies [937086b]
+  - @aws-amplify/cli-core@0.5.0-beta.2
+  - @aws-amplify/platform-core@0.5.0-beta.2
+
+## 0.7.0-beta.3
+
+### Minor Changes
+
+- eec100e: Updates the create flow with colors, verbosity and better structure
+
+### Patch Changes
+
+- Updated dependencies [3998cd3]
+- Updated dependencies [8d9a7a4]
+- Updated dependencies [b0ba24d]
+- Updated dependencies [8d9a7a4]
+  - @aws-amplify/cli-core@0.5.0-beta.1
+
 ## 0.6.1-beta.2
 
 ### Patch Changes